requirements.txt
# UI Framework
streamlit>=1.30.0

# Core Framework (for reference)
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
markdown==3.5.2
unstructured==0.12.0

# Embeddings & LLM
sentence-transformers==2.3.1
transformers==4.37.0
tiktoken==0.5.2

# Vector Database
pinecone-client==3.0.2

# Utilities
numpy==1.26.3
pandas==2.2.0
python-dotenv==1.0.0
aiofiles==23.2.1
tenacity==8.2.3
groq>=1.0.0
